+++ app/src/Presets.tsx
@@ -1,0 +1,91 @@
+import { useState } from "react";
+import { useContext } from "react";
+import { observer } from "mobx-react-lite";
+import { NavigateFunction, useNavigate } from "react-router-dom";
+import { ChartConfig, DatapackConfigForChartRequest, assertDatapackConfigForChartRequest } from "@tsconline/shared";
+import { context } from "./state";
+import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
+import { Accordion, AccordionSummary, AccordionDetails, Grid, Typography, Box } from "@mui/material";
+import { useTheme } from "@mui/material/styles";
+import { TSCCard, StyledScrollbar } from "./components";
+import "./Home.css";
+import { ErrorCodes } from "./util/error-codes";
+import _ from "lodash";
+import { useTranslation } from "react-i18next";
+
+export const Presets = () => {
+  const { state } = useContext(context);
+  const navigate = useNavigate();
+  return (
+    <Box>
+      {Object.entries(state.presets).map(([type, configArray]) => {
+        return <TSCPresetHighlights key={type} navigate={navigate} configArray={configArray} type={type} />;
+      })}
+    </Box>
+  );
+};
+const TSCPresetHighlights = observer(function TSCPresetHighlights({
+  type,
+  navigate,
+  configArray
+}: {
+  type: string;
+  navigate: NavigateFunction;
+  configArray: ChartConfig[];
+}) {
+  const { actions, state } = useContext(context);
+  const theme = useTheme();
+  const [expanded, setExpanded] = useState(true);
+  const { t } = useTranslation();
+  const handleAccordionChange = () => {
+    setExpanded(!expanded);
+  };
+  return (
+    <>
+      <Accordion
+        className="preset-highlight"
+        sx={{ border: `1px solid ${theme.palette.divider}` }}
+        onChange={handleAccordionChange}
+        expanded={expanded}>
+        <AccordionSummary
+          expandIcon={<ExpandMoreIcon />}
+          aria-controls="panel1a-content"
+          id="panel1a-header"
+          className="preset-summary">
+          <Typography className="preset-type-title">{t(`presets.${type}`)}</Typography>
+        </AccordionSummary>
+        <AccordionDetails>
+          <StyledScrollbar>
+            <Grid className="presets" container>
+              {configArray.map((preset, index) => (
+                <Grid item key={index} className="preset-item">
+                  <TSCCard
+                    preset={preset}
+                    generateChart={async () => {
+                      let datapacks: DatapackConfigForChartRequest[] = [];
+                      try {
+                        datapacks = preset.datapacks.map((dp) => {
+                          const datapack = _.cloneDeep(
+                            state.datapacks.find((d) => d.title === dp.name && d.type == "official")
+                          );
+                          assertDatapackConfigForChartRequest(datapack);
+                          return datapack;
+                        });
+                      } catch (e) {
+                        actions.pushError(ErrorCodes.NO_DATAPACK_FILE_FOUND);
+                        return;
+                      }
+                      const success = await actions.processDatapackConfig(datapacks, preset.settings);
+                      if (!success) return;
+                      actions.initiateChartGeneration(navigate, "/home");
+                    }}
+                  />
+                </Grid>
+              ))}
+            </Grid>
+          </StyledScrollbar>
+        </AccordionDetails>
+      </Accordion>
+    </>
+  );
+});
